requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
PyPDF2==3.0.1
faiss-cpu==1.8.0
requests==2.31.0
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0
numpy==1.26.4
huggingface-hub>=0.23.0
